+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/projects/ProjectSamplesController.java
@@ -806,16 +806,12 @@ public class ProjectSamplesController {
 	 * @throws IOException if the exported file cannot be written
 	 */
 	@RequestMapping(value = "/projects/{projectId}/samples/export")
-	public void exportProjectSamplesTable(
-			@PathVariable Long projectId,
-			@RequestParam DataTablesExportTypes type,
+	public void exportProjectSamplesTable(@PathVariable Long projectId, @RequestParam DataTablesExportTypes type,
 			@DataTablesRequest DataTablesParams params,
 			@RequestParam(required = false, defaultValue = "") List<String> sampleNames,
 			@RequestParam(required = false, defaultValue = "") List<Long> associated,
-			UISampleFilter filter,
-			HttpServletRequest request,
-			HttpServletResponse response,
-			Locale locale) throws IOException {
+			@RequestParam(required = false, name = "ids[]") List<Long> sampleIds, UISampleFilter filter,
+			HttpServletRequest request, HttpServletResponse response, Locale locale) throws IOException {
 
 		Project project = projectService.read(projectId);
 		List<Project> projects = new ArrayList<>();
@@ -825,15 +821,27 @@ public class ProjectSamplesController {
 		}
 		projects.add(project);
 
-		final Page<ProjectSampleJoin> page = sampleService
-				.getFilteredSamplesForProjects(projects, sampleNames, filter.getName(), params.getSearchValue(), filter.getOrganism(), filter.getStartDate(),
-						filter.getEndDate(), 0, Integer.MAX_VALUE, params.getSort());
+		final Page<ProjectSampleJoin> page;
+		page = sampleService.getFilteredSamplesForProjects(projects, sampleNames, filter.getName(),
+				params.getSearchValue(), filter.getOrganism(), filter.getStartDate(), filter.getEndDate(), 0,
+				Integer.MAX_VALUE, params.getSort());
 
 		// Create DataTables representation of the page.
-		List<DTProjectSamples> models = new ArrayList<>();
-		for (ProjectSampleJoin psj : page.getContent()) {
-			models.add(buildProjectSampleDataTablesModel(psj, locale));
-		}
+		List<DTProjectSamples> models = page.getContent()
+				.stream()
+				.filter(join -> {
+					if (sampleIds != null) {
+						// If a list of specific samples to export check to see if the sample is in the list
+						return sampleIds.contains(join.getObject()
+								.getId());
+					} else {
+						// If no specific samples asked for, export them all
+						return true;
+					}
+				})
+				.map(join -> buildProjectSampleDataTablesModel(join, locale))
+				.collect(Collectors.toUnmodifiableList());
+
 		List<String> headers = models.get(0)
 				.getExportableTableHeaders(messageSource, locale);
 		DataTablesExportToFile.writeFile(type, response, project.getLabel()