+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changes
 * [Test]: Updated chromedriver to version 92.0.0.
 * [UI]: Updated Project Single Sample Analyses Outputs pages to use Ant Design.
 * [UI]: Updated Your Single Sample Analysis Outputs page to use Ant Design.
+* [UI] Fixed bug which prevented the first created metadata template from being displayed after creation.
 
 21.01 to 21.05
 --------------