+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Changes
 * [UI]: Fixed bug where the length of the project name was not checked on project creation.
 * [UI] Fixed bug which prevented the first created metadata template from being displayed after creation.
 * [UI]: Sample modified date is now updated when sample metadata is modified.
+* [Developer]: Fix dependabot warning for `set-value`, force using a set-value above 4.0.1
 
 21.01 to 21.05
 --------------