+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changes
 =======
 
+21.09 to 22.01
+--------------
+
 21.05 to 21.09
 --------------
 * [Developer]: Added pipeline plugin path configuration option